Update link script.

+++ bin/linkme.sh
@@ -1,13 +1,12 @@
 #!/bin/bash
 
 function relink() {
-  rm -i $1
+  if [-e $1] ; then rm -i $1; fi
   ln -sn $2 $1
 }
 
 cd ~
-    relink .bash_profile ~/.dotfiles/.bash_profile
-    relink .bashrc ~/.dotfiles/.bashrc
-    relink .gitconfig ~/.dotfiles/.gitconfig
-    relink bin ~/.dotfiles/bin
-popd
+relink .bash_profile ~/.dotfiles/.bash_profile
+relink .bashrc ~/.dotfiles/.bashrc
+relink .gitconfig ~/.dotfiles/.gitconfig
+relink bin ~/.dotfiles/bin
